Improve in-person main flow happy path test, remove unnecessary duplicated test parts, improve inHouseMedicationsPage tests
