+++ apps/ehr/tests/e2e/specs/in-person/inHouseMedicationsPage.spec.ts
@@ -1,9 +1,9 @@
-import { Page, test } from '@playwright/test';
+import { BrowserContext, Page, test } from '@playwright/test';
 import { DateTime } from 'luxon';
 import { InPersonHeader } from 'tests/e2e/page/InPersonHeader';
 import { SideMenu } from 'tests/e2e/page/SideMenu';
 import { ResourceHandler } from '../../../e2e-utils/resource-handler';
-import { Field } from '../../page/EditMedicationCard';
+import { DIAGNOSIS_EMPTY_VALUE, Field, ORDERED_BY_EMPTY_VALUE } from '../../page/EditMedicationCard';
 import { InHouseMedicationsPage } from '../../page/in-person/InHouseMedicationsPage';
 import { expectAssessmentPage } from '../../page/in-person/InPersonAssessmentPage';
 import { expectEditOrderPage, OrderMedicationPage } from '../../page/OrderMedicationPage';
@@ -29,25 +29,61 @@ const NEW_ROUTE = 'Topical route';
 const NEW_INSTRUCTIONS = 'Edited instructions';
 const STATUS = 'pending';
 
-test.beforeEach(async () => {
+let page: Page;
+let context: BrowserContext;
+
+let orderMedicationPage: OrderMedicationPage;
+
+test.beforeAll(async ({ browser }) => {
+  context = await browser.newContext();
+  page = await context.newPage();
   if (process.env.INTEGRATION_TEST === 'true') {
     await resourceHandler.setResourcesFast();
   } else {
     await resourceHandler.setResources();
     await resourceHandler.waitTillAppointmentPreprocessed(resourceHandler.appointment.id!);
   }
+
+  await page.goto(`in-person/${resourceHandler.appointment.id}`);
+  const inPersonHeader = new InPersonHeader(page);
+  await inPersonHeader.selectIntakePractitioner();
+  await inPersonHeader.selectProviderPractitioner();
+  await inPersonHeader.clickSwitchModeButton('provider');
+  const sideMenu = new SideMenu(page);
+  await sideMenu.clickAssessment();
+  const assessmentPage = await expectAssessmentPage(page);
+  await assessmentPage.selectDiagnosis({ diagnosisNamePart: DIAGNOSIS });
+  await assessmentPage.selectDiagnosis({ diagnosisNamePart: NEW_DIAGNOSIS });
+  const inHouseMedicationsPage = await sideMenu.clickInHouseMedications();
+  orderMedicationPage = await inHouseMedicationsPage.clickOrderButton();
 });
 
-test.afterEach(async () => {
+test.afterAll(async () => {
   await resourceHandler.cleanupResources();
+  await page.close();
+  await context.close();
 });
 
-test('Open Order Medication screen, check required fields', async ({ page }) => {
-  const orderMedicationPage = await prepareAndOpenOrderMedicationPage(page);
+test('In-house medications page', async () => {
+  let editOrderPage: OrderMedicationPage;
+  let medicationsPage: InHouseMedicationsPage;
+
+  await test.step('"Order" button is disabled when all fields are empty', async () => {
+    await orderMedicationPage.editMedicationCard.selectAssociatedDx(DIAGNOSIS_EMPTY_VALUE);
+    await orderMedicationPage.editMedicationCard.selectOrderedBy(ORDERED_BY_EMPTY_VALUE);
+    await orderMedicationPage.verifyFillOrderToSaveButtonDisabled();
+  });
+  await test.step('Non-selected diagnosis on Assessment page is not present in Order Medication screen on associatedDx dropdown', async () => {
+    // cSpell:disable-next Loiasis
+    await orderMedicationPage.editMedicationCard.verifyDiagnosisNotAllowed('Loiasis');
+  });
+
+  // select diagnosis back
+  await orderMedicationPage.editMedicationCard.chooseOption(DIAGNOSIS);
   // we have selected dx by default now so we can proceed to verification
   await orderMedicationPage.clickOrderMedicationButton();
   await orderMedicationPage.editMedicationCard.verifyValidationErrorShown(Field.MEDICATION);
-  await orderMedicationPage.editMedicationCard.selectAssociatedDx('Select Associated Dx');
+  await orderMedicationPage.editMedicationCard.selectAssociatedDx(DIAGNOSIS_EMPTY_VALUE);
   await orderMedicationPage.editMedicationCard.selectMedication(MEDICATION);
   await orderMedicationPage.clickOrderMedicationButton();
   await orderMedicationPage.editMedicationCard.verifyValidationErrorShown(Field.DOSE);
@@ -58,120 +94,71 @@ test('Open Order Medication screen, check required fields', async ({ page }) => 
   await orderMedicationPage.clickOrderMedicationButton();
   await orderMedicationPage.editMedicationCard.verifyValidationErrorShown(Field.ROUTE);
   await orderMedicationPage.editMedicationCard.selectRoute(ROUTE);
+  await orderMedicationPage.editMedicationCard.waitForLoadOrderedBy();
+  await orderMedicationPage.editMedicationCard.selectFirstNonEmptyOrderedBy();
   await orderMedicationPage.clickOrderMedicationButton();
 
   await orderMedicationPage.editMedicationCard.verifyValidationErrorNotShown(Field.ASSOCIATED_DX);
   await orderMedicationPage.editMedicationCard.verifyValidationErrorNotShown(Field.MANUFACTURER);
   await orderMedicationPage.editMedicationCard.verifyValidationErrorNotShown(Field.INSTRUCTIONS);
-});
 
-test('"Order" button is disabled when all fields are empty', async ({ page }) => {
-  const orderMedicationPage = await prepareAndOpenOrderMedicationPage(page);
-  await orderMedicationPage.editMedicationCard.selectAssociatedDx('Select Associated Dx');
-  await orderMedicationPage.editMedicationCard.selectOrderedBy('Select Ordered By');
-  await orderMedicationPage.verifyFillOrderToSaveButtonDisabled();
-});
+  await orderMedicationPage.editMedicationCard.expectSaved();
 
-test('Non-selected diagnosis on Assessment page is not present in Order Medication screen on associatedDx dropdown', async ({
-  page,
-}) => {
-  const orderMedicationPage = await prepareAndOpenOrderMedicationPage(page);
-  // cSpell:disable-next Loiasis
-  await orderMedicationPage.editMedicationCard.verifyDiagnosisNotAllowed('Loiasis');
-});
-
-test('Non-numeric values can not be entered into "Dose" field', async ({ page }) => {
-  const orderMedicationPage = await prepareAndOpenOrderMedicationPage(page);
-  await orderMedicationPage.editMedicationCard.enterDose('abc1dfg');
-  await orderMedicationPage.editMedicationCard.verifyDose('1');
-});
-
-test('Order medication, order is submitted successfully and entered data are displayed correctly', async ({ page }) => {
-  const createOrderPage = await prepareAndOpenOrderMedicationPage(page);
-  await createOrderPage.editMedicationCard.selectAssociatedDx(DIAGNOSIS);
-  await createOrderPage.editMedicationCard.selectMedication(MEDICATION);
-  await createOrderPage.editMedicationCard.enterDose(DOSE);
-  await createOrderPage.editMedicationCard.selectUnits(UNITS);
-  await createOrderPage.editMedicationCard.enterManufacturer(MANUFACTURER);
-  await createOrderPage.editMedicationCard.selectRoute(ROUTE);
-  await createOrderPage.editMedicationCard.enterInstructions(INSTRUCTIONS);
-  await createOrderPage.editMedicationCard.waitForLoadOrderedBy();
-  await createOrderPage.clickOrderMedicationButton();
-
-  const editOrderPage = await expectEditOrderPage(page);
-  await editOrderPage.editMedicationCard.verifyAssociatedDx(DIAGNOSIS);
-  await editOrderPage.editMedicationCard.verifyMedication(MEDICATION);
-  await editOrderPage.editMedicationCard.verifyDose(DOSE);
-  await editOrderPage.editMedicationCard.verifyUnits(UNITS);
-  await editOrderPage.editMedicationCard.verifyManufacturer(MANUFACTURER);
-  // need to uncomment when https://github.com/masslight/ottehr/issues/3712 is fixed
-  //await editOrderPage.editMedicationCard.verifyRoute(ROUTE);
-  await editOrderPage.editMedicationCard.verifyInstructions(INSTRUCTIONS);
-
-  const medicationsPage = await editOrderPage.clickBackButton();
-  await medicationsPage.verifyMedicationPresent({
-    medicationName: MEDICATION,
-    dose: DOSE,
-    route: ROUTE,
-    instructions: INSTRUCTIONS,
-    status: STATUS,
+  await test.step('Non-numeric values can not be entered into "Dose" field', async () => {
+    await orderMedicationPage.editMedicationCard.enterDose('abc1dfg');
+    // this is fine since will be transfered as 1 and then after saving will be 1
+    await orderMedicationPage.editMedicationCard.verifyDose('01');
   });
-  await medicationsPage.clickMedicationDetailsTab();
-  await medicationsPage.medicationDetails().verifyAssociatedDx(DIAGNOSIS);
-  await medicationsPage.medicationDetails().verifyMedication(MEDICATION);
-  await medicationsPage.medicationDetails().verifyDose(DOSE);
-  await medicationsPage.medicationDetails().verifyUnits(UNITS);
-  await medicationsPage.medicationDetails().verifyManufacturer(MANUFACTURER);
-  await medicationsPage.medicationDetails().verifyRoute(ROUTE);
-  await medicationsPage.medicationDetails().verifyInstructions(INSTRUCTIONS);
-});
 
-test('Edit order page is opened after clicking on pencil icon for order in "pending" status', async ({ page }) => {
-  const createOrderPage = await prepareAndOpenOrderMedicationPage(page);
-  let editOrderPage: OrderMedicationPage;
-  let medicationsPage: InHouseMedicationsPage;
+  await test.step('Order medication, order is submitted successfully and entered data are displayed correctly in edit order page', async () => {
+    await orderMedicationPage.editMedicationCard.selectAssociatedDx(DIAGNOSIS);
+    await orderMedicationPage.editMedicationCard.selectMedication(MEDICATION);
+    await orderMedicationPage.editMedicationCard.enterDose(DOSE);
+    await orderMedicationPage.editMedicationCard.selectUnits(UNITS);
+    await orderMedicationPage.editMedicationCard.enterManufacturer(MANUFACTURER);
+    await orderMedicationPage.editMedicationCard.selectRoute(ROUTE);
+    await orderMedicationPage.editMedicationCard.enterInstructions(INSTRUCTIONS);
+    await orderMedicationPage.editMedicationCard.waitForLoadOrderedBy();
+    await orderMedicationPage.clickOrderMedicationButton();
+    await orderMedicationPage.editMedicationCard.expectSaved();
 
-  await test.step('Create order', async () => {
-    await createOrderPage.editMedicationCard.selectAssociatedDx(DIAGNOSIS);
-    await createOrderPage.editMedicationCard.selectMedication(MEDICATION);
-    await createOrderPage.editMedicationCard.enterDose(DOSE);
-    await createOrderPage.editMedicationCard.selectUnits(UNITS);
-    await createOrderPage.editMedicationCard.enterManufacturer(MANUFACTURER);
-    await createOrderPage.editMedicationCard.selectRoute(ROUTE);
-    await createOrderPage.editMedicationCard.enterInstructions(INSTRUCTIONS);
-    await createOrderPage.editMedicationCard.waitForLoadOrderedBy();
-    await createOrderPage.clickOrderMedicationButton();
     editOrderPage = await expectEditOrderPage(page);
+    await editOrderPage.editMedicationCard.verifyAssociatedDx(DIAGNOSIS);
+    await editOrderPage.editMedicationCard.verifyMedication(MEDICATION);
+    await editOrderPage.editMedicationCard.verifyDose('0' + DOSE);
+    await editOrderPage.editMedicationCard.verifyUnits(UNITS);
+    await editOrderPage.editMedicationCard.verifyManufacturer(MANUFACTURER);
+    // need to uncomment when https://github.com/masslight/ottehr/issues/3712 is fixed
+    //await editOrderPage.editMedicationCard.verifyRoute(ROUTE);
+    await editOrderPage.editMedicationCard.verifyInstructions(INSTRUCTIONS);
+  });
+
+  await test.step('Verify order data is displayed correctly in medication details tab', async () => {
+    medicationsPage = await editOrderPage.clickBackButton();
+    await medicationsPage.verifyMedicationPresent({
+      medicationName: MEDICATION,
+      dose: DOSE,
+      route: ROUTE,
+      instructions: INSTRUCTIONS,
+      status: STATUS,
+    });
+    await medicationsPage.clickMedicationDetailsTab();
+    await medicationsPage.medicationDetails().verifyAssociatedDx(DIAGNOSIS);
+    await medicationsPage.medicationDetails().verifyMedication(MEDICATION);
+    await medicationsPage.medicationDetails().verifyDose(DOSE);
+    await medicationsPage.medicationDetails().verifyUnits(UNITS);
+    await medicationsPage.medicationDetails().verifyManufacturer(MANUFACTURER);
+    await medicationsPage.medicationDetails().verifyRoute(ROUTE);
+    await medicationsPage.medicationDetails().verifyInstructions(INSTRUCTIONS);
   });
 
   await test.step('Click on pencil icon opens Edit order page', async () => {
-    const medicationsPage = await editOrderPage.clickBackButton();
+    await medicationsPage.clickMarTab();
     await medicationsPage.clickPencilIcon();
     editOrderPage = await expectEditOrderPage(page);
   });
 
-  await test.step('Update fields to empty values and click on [Save] - Validation errors appears', async () => {
-    await editOrderPage.editMedicationCard.clearMedication();
-    await editOrderPage.editMedicationCard.selectAssociatedDx('Select Associated Dx');
-    await editOrderPage.editMedicationCard.clearDose();
-    await editOrderPage.editMedicationCard.selectUnits('Select Units');
-    await editOrderPage.editMedicationCard.clearManufacturer();
-    await editOrderPage.editMedicationCard.clearInstructions();
-    await editOrderPage.clickOrderMedicationButton();
-
-    await editOrderPage.editMedicationCard.verifyValidationErrorShown(Field.MEDICATION);
-    await editOrderPage.editMedicationCard.verifyValidationErrorShown(Field.DOSE, false);
-    await editOrderPage.editMedicationCard.verifyValidationErrorShown(Field.UNITS, false);
-    await editOrderPage.editMedicationCard.verifyValidationErrorNotShown(Field.MANUFACTURER);
-    await editOrderPage.editMedicationCard.verifyValidationErrorNotShown(Field.INSTRUCTIONS);
-  });
-
-  await test.step('Edit order page is opened after clicking on pencil icon for order in "pending" status', async () => {
-    medicationsPage = await editOrderPage.clickBackButton();
-    await medicationsPage.clickPencilIcon();
-
-    //click on pencil icon opens Edit order page
-    editOrderPage = await expectEditOrderPage(page);
+  await test.step('Update order details and save, check on medications tabl', async () => {
     //Updated values are saved successfully and Order is updated on the "MAR" tab
     await editOrderPage.editMedicationCard.selectMedication(NEW_MEDICATION);
     await editOrderPage.editMedicationCard.enterDose(NEW_DOSE);
@@ -182,6 +169,7 @@ test('Edit order page is opened after clicking on pencil icon for order in "pend
     await editOrderPage.editMedicationCard.selectRoute(NEW_ROUTE);
     await editOrderPage.editMedicationCard.enterInstructions(NEW_INSTRUCTIONS);
     await editOrderPage.clickOrderMedicationButton();
+    await orderMedicationPage.editMedicationCard.expectSaved();
 
     await editOrderPage.editMedicationCard.verifyMedication(NEW_MEDICATION);
     await editOrderPage.editMedicationCard.verifyDose(NEW_DOSE);
@@ -200,18 +188,3 @@ test('Edit order page is opened after clicking on pencil icon for order in "pend
     });
   });
 });
-
-async function prepareAndOpenOrderMedicationPage(page: Page): Promise<OrderMedicationPage> {
-  await page.goto(`in-person/${resourceHandler.appointment.id}`);
-  const inPersonHeader = new InPersonHeader(page);
-  await inPersonHeader.selectIntakePractitioner();
-  await inPersonHeader.selectProviderPractitioner();
-  await inPersonHeader.clickSwitchModeButton('provider');
-  const sideMenu = new SideMenu(page);
-  await sideMenu.clickAssessment();
-  const assessmentPage = await expectAssessmentPage(page);
-  await assessmentPage.selectDiagnosis({ diagnosisNamePart: DIAGNOSIS });
-  await assessmentPage.selectDiagnosis({ diagnosisNamePart: NEW_DIAGNOSIS });
-  const inHouseMedicationsPage = await sideMenu.clickInHouseMedications();
-  return await inHouseMedicationsPage.clickOrderButton();
-}
