+++ apps/ehr/tests/e2e/specs/in-person/inPersonVisit.spec.ts
@@ -1,0 +1,196 @@
+import { BrowserContext, expect, Page, test } from '@playwright/test';
+import { QuestionnaireItemAnswerOption } from 'fhir/r4b';
+import { DateTime } from 'luxon';
+import { dataTestIds } from 'src/constants/data-test-ids';
+import { InPersonHeader } from 'tests/e2e/page/InPersonHeader';
+import {
+  chooseJson,
+  getConsentStepAnswers,
+  getContactInformationAnswers,
+  getEmergencyContactStepAnswers,
+  getPatientDetailsStepAnswers,
+  getPaymentOptionInsuranceAnswers,
+  getResponsiblePartyStepAnswers,
+  INSURANCE_PLAN_PAYER_META_TAG_CODE,
+  isoToDateObject,
+} from 'utils';
+import { ENV_LOCATION_NAME } from '../../../e2e-utils/resource/constants';
+import {
+  PATIENT_INSURANCE_MEMBER_ID,
+  PATIENT_INSURANCE_PLAN_TYPE,
+  PATIENT_INSURANCE_PLAN_TYPE_2,
+  PATIENT_INSURANCE_POLICY_HOLDER_2_ADDRESS,
+  PATIENT_INSURANCE_POLICY_HOLDER_2_ADDRESS_ADDITIONAL_LINE,
+  PATIENT_INSURANCE_POLICY_HOLDER_2_ADDRESS_AS_PATIENT,
+  PATIENT_INSURANCE_POLICY_HOLDER_2_BIRTH_SEX,
+  PATIENT_INSURANCE_POLICY_HOLDER_2_CITY,
+  PATIENT_INSURANCE_POLICY_HOLDER_2_DATE_OF_BIRTH,
+  PATIENT_INSURANCE_POLICY_HOLDER_2_FIRST_NAME,
+  PATIENT_INSURANCE_POLICY_HOLDER_2_LAST_NAME,
+  PATIENT_INSURANCE_POLICY_HOLDER_2_MIDDLE_NAME,
+  PATIENT_INSURANCE_POLICY_HOLDER_2_RELATIONSHIP_TO_INSURED,
+  PATIENT_INSURANCE_POLICY_HOLDER_2_STATE,
+  PATIENT_INSURANCE_POLICY_HOLDER_2_ZIP,
+  PATIENT_INSURANCE_POLICY_HOLDER_ADDRESS,
+  PATIENT_INSURANCE_POLICY_HOLDER_ADDRESS_ADDITIONAL_LINE,
+  PATIENT_INSURANCE_POLICY_HOLDER_ADDRESS_AS_PATIENT,
+  PATIENT_INSURANCE_POLICY_HOLDER_BIRTH_SEX,
+  PATIENT_INSURANCE_POLICY_HOLDER_CITY,
+  PATIENT_INSURANCE_POLICY_HOLDER_DATE_OF_BIRTH,
+  PATIENT_INSURANCE_POLICY_HOLDER_FIRST_NAME,
+  PATIENT_INSURANCE_POLICY_HOLDER_LAST_NAME,
+  PATIENT_INSURANCE_POLICY_HOLDER_MIDDLE_NAME,
+  PATIENT_INSURANCE_POLICY_HOLDER_RELATIONSHIP_TO_INSURED,
+  PATIENT_INSURANCE_POLICY_HOLDER_STATE,
+  PATIENT_INSURANCE_POLICY_HOLDER_ZIP,
+  ResourceHandler,
+} from '../../../e2e-utils/resource-handler';
+import { expectAssessmentPage } from '../../page/in-person/InPersonAssessmentPage';
+import { expectInPersonProgressNotePage } from '../../page/in-person/InPersonProgressNotePage';
+import { expectPatientInfoPage } from '../../page/PatientInfo';
+import { openVisitsPage } from '../../page/VisitsPage';
+
+// cSpell:disable-next inversus
+const DIAGNOSIS = 'Situs inversus';
+const EM_CODE = '99201 New Patient - E/M Level 1';
+
+test.describe('In-person visit', async () => {
+  // test('Book appointment, go to Hospitalization page and complete Intake, check statuses', async ({ page }) => {
+  //   const patientInfoPage = await intakeTestAppointment(page, resourceHandler);
+  //   const hospitalizationPage = await patientInfoPage.sideMenu().clickHospitalization();
+  //   await hospitalizationPage.clickCompleteIntakeButton();
+  //   await patientInfoPage.inPersonHeader().verifyStatus('ready for provider');
+  // });
+
+  test.describe('happy path', () => {
+    const PROCESS_ID = `inPersonVisit.spec.ts-${DateTime.now().toMillis()}`;
+    let insuranceCarrier1: QuestionnaireItemAnswerOption | undefined;
+    let insuranceCarrier2: QuestionnaireItemAnswerOption | undefined;
+
+    const resourceHandler = new ResourceHandler(PROCESS_ID, 'in-person', async ({ patientInfo }) => {
+      return [
+        getContactInformationAnswers({
+          firstName: patientInfo.firstName,
+          lastName: patientInfo.lastName,
+          birthDate: isoToDateObject(patientInfo.dateOfBirth || '') || undefined,
+          email: patientInfo.email,
+          phoneNumber: patientInfo.phoneNumber,
+          birthSex: patientInfo.sex,
+        }),
+        getPatientDetailsStepAnswers({}),
+        getPaymentOptionInsuranceAnswers({
+          insuranceCarrier: insuranceCarrier1!,
+          insurancePlanType: PATIENT_INSURANCE_PLAN_TYPE,
+          insuranceMemberId: PATIENT_INSURANCE_MEMBER_ID,
+          insurancePolicyHolderFirstName: PATIENT_INSURANCE_POLICY_HOLDER_FIRST_NAME,
+          insurancePolicyHolderLastName: PATIENT_INSURANCE_POLICY_HOLDER_LAST_NAME,
+          insurancePolicyHolderMiddleName: PATIENT_INSURANCE_POLICY_HOLDER_MIDDLE_NAME,
+          insurancePolicyHolderDateOfBirth: PATIENT_INSURANCE_POLICY_HOLDER_DATE_OF_BIRTH,
+          insurancePolicyHolderBirthSex: PATIENT_INSURANCE_POLICY_HOLDER_BIRTH_SEX,
+          insurancePolicyHolderAddressAsPatient: PATIENT_INSURANCE_POLICY_HOLDER_ADDRESS_AS_PATIENT,
+          insurancePolicyHolderAddress: PATIENT_INSURANCE_POLICY_HOLDER_ADDRESS,
+          insurancePolicyHolderAddressAdditionalLine: PATIENT_INSURANCE_POLICY_HOLDER_ADDRESS_ADDITIONAL_LINE,
+          insurancePolicyHolderCity: PATIENT_INSURANCE_POLICY_HOLDER_CITY,
+          insurancePolicyHolderState: PATIENT_INSURANCE_POLICY_HOLDER_STATE,
+          insurancePolicyHolderZip: PATIENT_INSURANCE_POLICY_HOLDER_ZIP,
+          insurancePolicyHolderRelationshipToInsured: PATIENT_INSURANCE_POLICY_HOLDER_RELATIONSHIP_TO_INSURED,
+          insuranceCarrier2: insuranceCarrier2!,
+          insurancePlanType2: PATIENT_INSURANCE_PLAN_TYPE_2,
+          insuranceMemberId2: PATIENT_INSURANCE_MEMBER_ID,
+          insurancePolicyHolderFirstName2: PATIENT_INSURANCE_POLICY_HOLDER_2_FIRST_NAME,
+          insurancePolicyHolderLastName2: PATIENT_INSURANCE_POLICY_HOLDER_2_LAST_NAME,
+          insurancePolicyHolderMiddleName2: PATIENT_INSURANCE_POLICY_HOLDER_2_MIDDLE_NAME,
+          insurancePolicyHolderDateOfBirth2: PATIENT_INSURANCE_POLICY_HOLDER_2_DATE_OF_BIRTH,
+          insurancePolicyHolderBirthSex2: PATIENT_INSURANCE_POLICY_HOLDER_2_BIRTH_SEX,
+          insurancePolicyHolderAddressAsPatient2: PATIENT_INSURANCE_POLICY_HOLDER_2_ADDRESS_AS_PATIENT,
+          insurancePolicyHolderAddress2: PATIENT_INSURANCE_POLICY_HOLDER_2_ADDRESS,
+          insurancePolicyHolderAddressAdditionalLine2: PATIENT_INSURANCE_POLICY_HOLDER_2_ADDRESS_ADDITIONAL_LINE,
+          insurancePolicyHolderCity2: PATIENT_INSURANCE_POLICY_HOLDER_2_CITY,
+          insurancePolicyHolderState2: PATIENT_INSURANCE_POLICY_HOLDER_2_STATE,
+          insurancePolicyHolderZip2: PATIENT_INSURANCE_POLICY_HOLDER_2_ZIP,
+          insurancePolicyHolderRelationshipToInsured2: PATIENT_INSURANCE_POLICY_HOLDER_2_RELATIONSHIP_TO_INSURED,
+        }),
+        getResponsiblePartyStepAnswers({}),
+        getEmergencyContactStepAnswers({}),
+        getConsentStepAnswers({}),
+      ];
+    });
+    let page: Page;
+    let context: BrowserContext;
+
+    test.beforeAll(async ({ browser }) => {
+      const oystehr = await ResourceHandler.getOystehr();
+      const insuranceCarriersOptionsResponse = await oystehr.zambda.execute({
+        id: 'get-answer-options',
+        answerSource: {
+          resourceType: 'InsurancePlan',
+          query: `status=active&_tag=${INSURANCE_PLAN_PAYER_META_TAG_CODE}`,
+        },
+      });
+      const insuranceCarriersOptions = chooseJson(insuranceCarriersOptionsResponse) as QuestionnaireItemAnswerOption[];
+
+      insuranceCarrier1 = insuranceCarriersOptions.at(0);
+      insuranceCarrier2 = insuranceCarriersOptions.at(1);
+      if (process.env.INTEGRATION_TEST === 'true') {
+        await resourceHandler.setResourcesFast();
+      } else {
+        await resourceHandler.setResources();
+        await resourceHandler.waitTillHarvestingDone(resourceHandler.appointment.id!);
+        await resourceHandler.waitTillAppointmentPreprocessed(resourceHandler.appointment.id!);
+      }
+
+      context = await browser.newContext();
+      page = await context.newPage();
+    });
+
+    test.afterAll(async () => {
+      await resourceHandler.cleanupResources();
+      await page.close();
+      await context.close();
+    });
+
+    test('Create appointment, tracking board change statuses, fill required info and complete visit', async ({
+      page,
+    }) => {
+      test.setTimeout(240000); // 4 minutes
+
+      const visitsPage = await openVisitsPage(page);
+      await visitsPage.selectLocation(ENV_LOCATION_NAME!);
+      await visitsPage.clickPrebookedTab();
+      await visitsPage.clickArrivedButton(resourceHandler.appointment.id!);
+      await visitsPage.clickInOfficeTab();
+      await visitsPage.verifyVisitsStatus(resourceHandler.appointment.id!, 'arrived');
+      await visitsPage.clickIntakeButton(resourceHandler.appointment.id!);
+      await expect(async () => {
+        const content = await page.getByTestId(dataTestIds.inPersonHeader.container).textContent();
+        return content?.includes(resourceHandler.patient.name![0].family!) ?? false;
+      }).toPass({ timeout: 30_000 });
+      const inPersonHeader = new InPersonHeader(page);
+      await inPersonHeader.selectIntakePractitioner();
+      await inPersonHeader.selectProviderPractitioner();
+      const patientInfoPage = await expectPatientInfoPage(resourceHandler.appointment.id!, page);
+
+      await patientInfoPage.inPersonHeader().verifyStatus('intake');
+      await patientInfoPage.sideMenu().clickCompleteIntakeButton();
+      await patientInfoPage.inPersonHeader().verifyStatus('ready for provider');
+      await patientInfoPage.inPersonHeader().clickSwitchModeButton('provider');
+      await patientInfoPage.inPersonHeader().changeStatus('provider');
+      const progressNotePage = await expectInPersonProgressNotePage(page);
+      await progressNotePage.verifyReviewAndSignButtonDisabled();
+      await progressNotePage.fillHPI();
+      await patientInfoPage.sideMenu().clickAssessment();
+      const assessmentPage = await expectAssessmentPage(page);
+      await assessmentPage.selectDiagnosis({ diagnosisNamePart: DIAGNOSIS });
+      await assessmentPage.selectEmCode(EM_CODE);
+      await patientInfoPage.sideMenu().clickProgressNote();
+      await progressNotePage.clickDischargeButton();
+      await progressNotePage.clickReviewAndSignButton();
+      await progressNotePage.clickSignButton();
+      await patientInfoPage.inPersonHeader().verifyStatus('completed');
+      await openVisitsPage(page);
+      await visitsPage.selectLocation(ENV_LOCATION_NAME!);
+      await visitsPage.clickDischargedTab();
+      await visitsPage.verifyVisitPresent(resourceHandler.appointment.id!);
+    });
+  });
+});
